pyproject.toml
[build-system]
requires = ["hatchling>=1.13"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/edu"]

[project]
name = "fastapi-edu-cha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "bcrypt>=4.3.0",
    "dishka>=1.6.0",
    "fastapi>=0.116.1",
    "orjson>=3.11.3",
    "psycopg[binary]>=3.2.10",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy[asyncio]>=2.0.43",
    "uuid7>=0.1.0",
    "uvicorn>=0.35.0",
]
